# Optimize: skip setting zero during slice init